MAINT: Move plotting to _plot.py
