+++ sourcetracker/__init__.py
@@ -7,8 +7,10 @@
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
 
-from sourcetracker._compare import compare_sinks, compare_sink_metrics
-from sourcetracker._sourcetracker import gibbs, plot_heatmap
+from ._compare import compare_sinks, compare_sink_metrics
+from ._sourcetracker import gibbs
+from ._plot import plot_heatmap
+
 
 __version__ = '2.0.1-dev'
 _readme_url = "https://github.com/biota/sourcetracker2/blob/master/README.md"
