+++ sourcetracker/_cli/__init__.py
@@ -16,4 +16,5 @@ from sourcetracker import __version__
 def cli():
     pass
 
+
 import_module('sourcetracker._cli.gibbs')
