+++ sourcetracker/_sourcetracker.py
@@ -9,16 +9,10 @@
 # ----------------------------------------------------------------------------
 from __future__ import division
 
-import matplotlib
-matplotlib.use('Agg')  # noqa: E402
-
 import numpy as np
 from skbio.stats import subsample_counts
 import pandas as pd
 from functools import partial
-
-import seaborn as sns
-import matplotlib.pyplot as plt
 
 
 def validate_gibbs_input(sources, sinks=None):
@@ -1014,16 +1008,3 @@ def collate_gibbs_results(all_envcounts, all_env_assignments,
             fts = None
 
     return props, props_stds, fts
-
-
-def plot_heatmap(mpm, cm=plt.cm.viridis, xlabel='Sources', ylabel='Sinks',
-                 title='Mixing Proportions (as Fraction)'):
-    '''Make a basic mixing proportion histogram.'''
-    fig = plt.figure()
-    ax = fig.add_subplot(1, 1, 1)
-    sns.heatmap(mpm, vmin=0, vmax=1.0, cmap=cm, annot=True, linewidths=.5,
-                ax=ax)
-    ax.set_xlabel(xlabel)
-    ax.set_ylabel(ylabel)
-    ax.set_title(title)
-    return fig, ax
